requirements.txt
backcall==0.1.0
biopython==1.76
cycler==0.10.0
decorator==4.4.2
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.2
numpy==1.18.3
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
